openssh: delete key exchange algorithms
Reverts commit d7e889bf6d866ad90ce7bf243155b79cf2fa4600

The KexAlgorithms set do not meaningfully improve security beyond what
the current nixpkgs ships by default. The difference between the two is
the following, in a pseudo-`diff nixpkgs..srvos` format:

-mlkem768x25519-sha256
-sntrup761x25519-sha512
+diffie-hellman-group16-sha512
+diffie-hellman-group18-sha512
-diffie-hellman-group-exchange-sha256

The first two algorithms are modern post-quantum algorithms on top of
25519, generally regarded as secure.

The last three algorithms are also regarded as secure, though the last,
DH-group-exchange, depends on the shipped moduli files from nixpkgs.

Holding this configuration in srvos may be an unnecesary maintenance
burden, as these settings are not in any sense merged with the upstream
nixpkgs, and override it instead.

Regarding breaking ssh connections for users, I find it unlikely that there
are people depending on DH-group{16,18} who also support neither
DH-group-exchange nor curve25519

An alternative is to ship a blacklist in srvos, which simply requires
prepending a dash to the first element of the list. This still overrides
the nixpkgs defaults but fails open to accept new openSSH algorithms.